Add essential structures for multiblock support
Multiblocks are self contained, each block will be responsible for keeping track of whether or not it is in a constructed multiblock. IMultiblockComponent will be responsible for providing an interface to check for and locate mutliblock components and controllers.

MultiblockBlock is used once per multiblock structure and is responsible for constructed and destructing multiblocks, as well as storing their data. All ticking and processing will be done by its corresponding MultiblockParentBlockEntity. All blocks contained within a multiblock structure that are not its controller will have a MultiblockChildBlockEntity and will only be keeping track of where its controller is (if it is constructed and thus has one).