Add Arc Furnace multiblock structure
Added the Arc Furnance structure and some of its components. The plan is for it to be used to convert more valuable metal ores into ingots and iron into steel. As well as the overall structure, the EnergyPortBlock has been added, to feed energy into the multiblock structure.

In addition to the EnergyPortBlock block, some form of item interface will be needed to automate the transfer of ingredients and products, as well as a recipe handler for the structure.