Add Waila support
Added Waila support for multiblocks. The base MultiblockBlock now has a "Multiblock" indicator that shows whether or not the structure has been constructed. Subsequent multiblock components now show the tooltip for the master MultiblockBlock. In order to better suit the new tooltip, the ArcFurnaceControllerBlock and related classes were renamed to ArcFurnaceBlock and so on.

With improved integration, the tooltip could include the structure's capabilities when hovering over any block of the overall structure.