Add universe related data
Added universe data classes and the networking necessary to send that data to the client. Because of the universal nature of the planet data, the server stores an instance of SavedData in the overworld to hold the planet data globally. Upon joining a level (not just on log in), a UniversePacket is sent to the client. The data within this packet is stored locally on a capability that is only attached to client-side levels.

The universe data can be accessed from Universe#getInstance. This
function will automatically either load the SavedData, if on the server,
or get the local level's capability data.